+++ src/task/dto/task.go
@@ -13,7 +13,7 @@ import (
 type CallTask struct {
 	TaskCodeName string
 
-	Timeout   *int64 `json:"timeout" valid:"Range(0,86400000)"`
+	Timeout   *int32 `json:"timeout" valid:"Range(0,86400000)"`
 	Arguments string `json:"arguments" valid:"Required;MinSize(2)"`
 }
